Locked Pets 0.1
Doesnt fully work yet! Currently having problems with how the data works n jquery stuff. Temporarily made doggo_1 and 4 locked hovers because they spam the console for some reason.

What works: Manually making one true in the progress will unlock the dog. False dogs are locked.

Bug: Locked pets don't show static text.

To be implemented: Unlocking, though this is part with the puzzle, I was trying to implement a modal for the mean time to simulate it (need help).